Allow watching extensionless files (e.g. Dockerfile) (#14)
This PR updates the MatchesExts function to allow watched extensions to match entire filenames
if those files don't have an additional extension (e.g. Dockerfile).

This could cause weird edge cases (e.g. if somebody creates a file called js rather than something.js
then it would be synced. But if they are doing that, then it might be useful for the coach to know about
anyway, and is unlikely to be a huge problem I think.

---

It also allows leading periods in the specified extensions (e.g. `.js .ts`). This will allow a migration 
to always specifying `.`, and removing any ambiguity between e.g. Dockerfile and .Dockerfile