Allocate random available port per test
Previously, tests were failing for the second test to use the mocked
server. A time.Sleep solved this, so my theory is that the previous
server shutdown was interfering in some way with the setup of the new
one. By allocating a different random available port each time we will
hopefully solve this issue.