Allocate different available port per test (#16)
Previously, requests in tests were failing for the second test to use the mocked
server. A time.Sleep solved this, and reordering tests always resulted in the
second one failing. We therefore theorised that the previous server shutdown 
was interfering in some way with the setup of the new one. 

A possible explanation is Linux holding the old listener in a TIME_WAIT state.  

Allocating a different random available port each time appears to solve this issue.

Another potential solution (which is not implemented in this PR) would be to set
SO_REUSEADDR on the TCP listener.