Disable the http endpoint in the hosted env
The http endpoint is an alternative to supplying an attendance id file / env variable.
The idea was that the user's browser could tell sync what the id is.

In the hosted env this doesn't work without extra steps (since this assumes
that the browser and the sync process can talk to each other) moreover it means
that running multiple sync processes in different directories fails on startup.

Therefore this commit disables the http endpoint when the hosted env is
detected