Handle retry errors properly (#28)
Previously we treated all 4xx errors as client errors and just exits. While this might be valid for errors that are synonymous with misconfiguration (eg bad attendance id),
this is not the correct behaviour for a 429 response, which is what our server side rate limiter returns. Especially confusing when sync is running in a docker container
and its output is not visible unless you go looking for it. In this case we should just continue with our retry / backoff strategy.

In a simulated example (a script updating files over and over) this seems to behave as intended: we'll back off and retry & eventually the updates get through.